git-svn-id: https://subversion.assembla.com/svn/kbasm/cpgf/trunk@564 5ea4cdcf-77fc-412b-903d-47c6f9a32d48

+++ include/cpgf/private/gmetaoperator_p.h
@@ -657,7 +657,7 @@ private:
 	}
 
 	static GMetaType virtualGetParamType(size_t index) {
-		if(index < static_cast<size_t>(FT::Arity)) {
+		if((int)index < static_cast<int>(FT::Arity)) {
 			switch(index) {
 #define REF_GETPARAM_HELPER(N, unused) \
 	case N: return createMetaType<typename TypeList_GetWithDefault<typename FT::ArgTypeList, N>::Result>();
